pyproject.toml
[tool.poetry]
name = "status-logger"
version = "0.1.0"
description = "Client-server script which logs the status count."
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
pymongo = "^4.8.0"
paho-mqtt = "^2.1.0"
ruff = "^0.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
